* Merged the branch trinity to the trunk (r87:159)    Please DONT use the trinity branch anymore.

+++ common/RowsetReader.cpp
@@ -203,7 +203,7 @@ RowsetReader::iterator RowsetReader::begin() {
 	return(iterator(this, false));
 }
 
-uint32 RowsetReader::size() const {
+size_t RowsetReader::size() const {
 	return(m_set->lines.items.size());
 }
